refactor(final form coponents): rename folder name
