+++ src/components/Inputs/DataElementIdScheme.js
@@ -1,6 +1,6 @@
 import i18n from '@dhis2/d2-i18n'
 import React from 'react'
-import { RadioGroup } from '../form/RadioGroup'
+import { RadioGroup } from '../FinalFormComponents/RadioGroup'
 
 export const OPTION_UID = { value: 'UID', label: i18n.t('Uid') }
 export const OPTION_CODE = { value: 'CODE', label: i18n.t('Code') }
