+++ src/components/Inputs/DryRun.js
@@ -1,6 +1,6 @@
 import i18n from '@dhis2/d2-i18n'
 import React from 'react'
-import { RadioGroup } from '../form/RadioGroup'
+import { RadioGroup } from '../FinalFormComponents/RadioGroup'
 
 export const OPTION_YES = { value: 'true', label: i18n.t('Yes') }
 export const OPTION_NO = { value: 'false', label: i18n.t('No') }
