+++ src/components/Inputs/Strategy.js
@@ -1,6 +1,6 @@
 import i18n from '@dhis2/d2-i18n'
 import React from 'react'
-import { RadioGroup } from '../form/RadioGroup'
+import { RadioGroup } from '../FinalFormComponents/RadioGroup'
 
 export const OPTION_NEW_AND_UPDATES = {
     value: 'NEW_AND_UPDATES',
