+++ src/pages/import/Data.js
@@ -3,13 +3,13 @@ import React, { useState } from 'react'
 import cx from 'classnames'
 import i18n from '@dhis2/d2-i18n'
 
-import { Form, FormSpy } from 'react-final-form'
+import { Form } from 'react-final-form'
 
 import { DataElementIdScheme } from '../../components/Inputs/DataElementIdScheme'
 import { DataIcon } from '../../components/Icon'
 import { DryRun } from '../../components/Inputs/DryRun'
 import { Error } from '../../components/Error'
-import { File } from '../../components/form/File'
+import { File } from '../../components/FinalFormComponents/File'
 import { FormContent } from '../../components/FormContent'
 import { FormFoot } from '../../components/FormFoot'
 import { FormHead } from '../../components/FormHead'
